Added Normal Mapping and Skeletal Animation
Added normal mapping - can supply a tangent-space normal map on slot 2, which is
applied to change the world-space normals. Tangents & Bitangents are generated by Assimp.
Added skeletal animation - each bone is given a gameobject & transform. When a skinned mesh renders,
it looks up its root Animation Component, from which it can access the transforms of the bone
gameobjects, which it uses to get the bone matrices used in the shader.

Still needs a little more work (e.g. handling multiple animations), but for a single animation it works pretty well
Also have to deal with the inbound merge conflicts ;)