Quick fix for deferred boids
Quick fix so boids use the deferred shader.
Also made a container object for the boids, which made it so boid rotation no longer needed a magic pi/2
to convert from blender axes to opengl axes

Still a problem with using the ANIM shader with boids - I realized it is because all changes to the root gameobject
need to not be applied to bones in the shader.